chore: organize project
