+++ lib/components/icons/ArrowBackIcon.ts
