+++ lib/components/icons/ErrorIcon.ts
