+++ lib/components/index.ts
@@ -1,3 +1,4 @@
 export * from './icons';
 export * from './layout';
+export * from './portfolio';
 export * from './presentational';
