+++ lib/components/portfolio/Charts/components/BarChartContainer/BarChartContainer.tsx
