+++ lib/components/portfolio/Charts/components/BarChartContainer/components/BarChart/BarChart.tsx
