+++ lib/components/portfolio/Charts/components/BarChartContainer/components/BarChart/index.ts
