+++ lib/components/portfolio/Charts/components/BarChartContainer/components/EmptyBarChart/EmptyBarChart.tsx
