+++ lib/components/portfolio/Charts/components/BarChartContainer/components/EmptyBarChart/index.ts
