+++ lib/components/portfolio/Charts/components/BarChartContainer/index.ts
