+++ lib/components/portfolio/Charts/components/Legend/Legend.tsx
