+++ lib/components/portfolio/Charts/components/LoadingChart/LoadingChart.tsx
