+++ lib/components/portfolio/Charts/components/LoadingChart/index.ts
